Thu, Jul 27, 2023, 2:43 AM -03:00

+++ caesar/caesar.c
@@ -19,7 +19,7 @@ int main(int argc, char *argv[])
 
         if(keyCheckFunc(argv[1]) == 1) // check argv[1,i] = int
         {
-            key[0] = argv[1,0];
+            key[0] = (int)argv[1,0];
             printf("plaintext: ");
             scanf("%s", txt);
             for(i = 0 , len=strlen(txt) ; i < len ; i++)
